requirements.txt
asgiref==3.8.1
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
flake8==7.2.0
isort==6.0.1
mccabe==0.7.0
pillow==11.1.0
psycopg2==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
sqlparse==0.5.3
tzdata==2025.2
